requirements.txt
# Core dependencies
streamlit==1.38.0
pandas==2.2.2

# Email validation & domain checks
email-validator==2.2.0
dnspython==2.6.1

# String similarity for typo suggestions
python-Levenshtein==0.25.1

# Extra (UI/formatting)


# For CSV handling (optional, but keeps things consistent)
pyarrow==17.0.0

streamlit==1.38.0

email-validator==2.2.0
dnspython==2.6.1
rapidfuzz==3.9.6
rich==13.7.1
python-dotenv==1.0.1
